+++ js/main.js
@@ -1,13 +1,9 @@
 import {
   createOffersMarkup
 } from './offer.js';
+import './form.js';
 import {
-  resetForm,
-  resetButton
-} from './form.js';
-import {
-  addOfferMap,
-  resetFilter
+  addOfferMap
 } from './map.js';
 import {
   getData
@@ -19,10 +15,3 @@ const renderMarker = (data) => {
 };
 
 getData(renderMarker);
-
-const onResetButtonClick = (evt) => {
-  resetForm(evt);
-  resetFilter();
-};
-resetButton.addEventListener('click', onResetButtonClick)
-
